feat: 非同期で同時に DB に書き込む時に SQLite だと SQLITE_BUSY のエラーが発生することが多かったが H2 の方が同時書き込みに強いという話を聞いたので DB を SQLite から H2 に切り替える

+++ settings/src/main/scala/infrastructure/Settings.scala
@@ -10,7 +10,7 @@ object Settings {
 }
 
 trait SettingsType {
-  val watches: Map[Int, WatchSettings]
+  val watches: Map[Byte, WatchSettings]
   val slackToken: String
 }
